+++ src/crypt/aes_encrypt.rs
@@ -1,3 +1,4 @@
+use crate::storage::jenv::get_required_env_var;
 use aes_gcm::AeadCore;
 use aes_gcm::{
     aead::{Aead, KeyInit, OsRng},
@@ -5,11 +6,11 @@ use aes_gcm::{
 };
 use rand::{thread_rng, Rng};
 use sha2::{Digest, Sha256};
-use std::env;
 
 // Encrypt a private key with a password
 pub fn encrypt_private_key(private_key: &[u8]) -> Result<Vec<u8>, Box<dyn std::error::Error>> {
-    let password = env::var("JACS_PRIVATE_KEY_PASSWORD".to_string())?;
+    // Password is required but can be empty (false)
+    let password = get_required_env_var("JACS_PRIVATE_KEY_PASSWORD", false)?;
 
     // Generate a random salt
     let mut salt = [0u8; 16];
@@ -47,7 +48,8 @@ pub fn encrypt_private_key(private_key: &[u8]) -> Result<Vec<u8>, Box<dyn std::e
 pub fn decrypt_private_key(
     encrypted_key_with_salt_and_nonce: &[u8],
 ) -> Result<Vec<u8>, Box<dyn std::error::Error>> {
-    let password = env::var("JACS_PRIVATE_KEY_PASSWORD".to_string())?;
+    // Password is required but can be empty (false)
+    let password = get_required_env_var("JACS_PRIVATE_KEY_PASSWORD", false)?;
 
     if encrypted_key_with_salt_and_nonce.len() < 16 + 12 {
         return Err("encrypted data is too short".into());